backend/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
bidict==0.22.0
cachelib==0.9.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cmake==3.26.3
cryptography==40.0.2
cs50==9.2.5
dnspython==1.16.0
eventlet==0.30.2
Flask==2.2.5
Flask-Cors==3.0.10
Flask-Login==0.6.2
Flask-Session==0.4.0
Flask-SocketIO==5.3.4
frozenlist==1.3.3
greenlet==2.0.2
gunicorn==20.1.0
idna==3.4
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
jwt==1.3.1
MarkupSafe==2.1.1
multidict==6.0.4
mysqlclient==2.1.1
pika==1.3.2
pycparser==2.21
PyJWT==2.7.0
python-engineio==4.4.1
python-socketio==5.8.0
requests==2.31.0
six==1.11.0
SQLAlchemy==1.4.46
sqlparse==0.4.4
termcolor==2.3.0
urllib3==2.0.2
Werkzeug==2.2.3
yarl==1.9.2
zipp==3.8.1
